+++ driven/storage/adapter.go
@@ -454,7 +454,7 @@ func (sa Adapter) DeleteUserWithID(orgID string, appID string, userID string) er
 				return err
 			}
 
-			messages, err := sa.GetMessages(orgID, appID, &userID, nil, nil, nil, nil, nil, nil, nil)
+			messages, err := sa.GetMessages(orgID, appID, &userID, nil, nil, nil, nil, nil, nil, nil, nil, nil)
 			if err != nil {
 				fmt.Printf("warning: unable to retrieve messages for user (%s): %s\n", userID, err)
 				abortTransaction(sessionContext)
@@ -772,7 +772,7 @@ func (sa Adapter) UpdateTopic(topic *model.Topic) (*model.Topic, error) {
 }
 
 // GetMessages Gets all messages according to the filter
-func (sa Adapter) GetMessages(orgID string, appID string, userID *string, messageIDs []string, startDateEpoch *int64, endDateEpoch *int64, filterTopic *string, offset *int64, limit *int64, order *string) ([]model.Message, error) {
+func (sa Adapter) GetMessages(orgID string, appID string, userID *string, read *bool, mute *bool, messageIDs []string, startDateEpoch *int64, endDateEpoch *int64, filterTopic *string, offset *int64, limit *int64, order *string) ([]model.Message, error) {
 	filter := bson.D{
 		primitive.E{Key: "org_id", Value: orgID},
 		primitive.E{Key: "app_id", Value: appID},
@@ -781,6 +781,7 @@ func (sa Adapter) GetMessages(orgID string, appID string, userID *string, messag
 	if userID != nil {
 		innerFilter = append(innerFilter, bson.D{primitive.E{Key: "user_id", Value: userID}})
 	}
+
 	if len(innerFilter) > 0 {
 		filter = append(filter, primitive.E{Key: "recipients", Value: bson.D{primitive.E{Key: "$elemMatch", Value: bson.D{primitive.E{Key: "$or", Value: innerFilter}}}}})
 	}
@@ -790,6 +791,25 @@ func (sa Adapter) GetMessages(orgID string, appID string, userID *string, messag
 	if len(messageIDs) > 0 {
 		filter = append(filter, primitive.E{Key: "_id", Value: bson.M{"$in": messageIDs}})
 	}
+
+	if read != nil {
+		if *read {
+			filter = append(filter, primitive.E{Key: "recipients.read", Value: true})
+		} else {
+			// support of existing records where "read" field is missing
+			filter = append(filter, primitive.E{Key: "recipients.read", Value: bson.M{"$ne": true}})
+		}
+	}
+
+	if mute != nil {
+		if *mute {
+			filter = append(filter, primitive.E{Key: "recipients.mute", Value: true})
+		} else {
+			// support of existing records where "mute" field is missing
+			filter = append(filter, primitive.E{Key: "recipients.mute", Value: bson.M{"$ne": true}})
+		}
+	}
+
 	if startDateEpoch != nil {
 		seconds := *startDateEpoch / 1000
 		timeValue := time.Unix(seconds, 0)
@@ -955,6 +975,27 @@ func (sa Adapter) DeleteMessageWithContext(ctx context.Context, orgID string, ap
 	}
 
 	return nil
+}
+
+// UpdateUnreadMessage updates a unread message in the recipients to read
+func (sa Adapter) UpdateUnreadMessage(ctx context.Context, orgID string, appID string, ID string, userID *string) (*model.Message, error) {
+	read := true
+	filter := bson.D{primitive.E{Key: "_id", Value: ID},
+		primitive.E{Key: "app_id", Value: appID},
+		primitive.E{Key: "org_id", Value: orgID},
+		primitive.E{Key: "recipients.user_id", Value: userID}}
+	update := bson.D{
+		primitive.E{Key: "$set", Value: bson.D{
+			primitive.E{Key: "recipients.$.read", Value: read},
+			primitive.E{Key: "recipients.$.date_updated", Value: time.Now().UTC()},
+		}},
+	}
+	_, err := sa.db.messages.UpdateManyWithContext(ctx, filter, update, nil)
+	if err != nil {
+		fmt.Println("warning: error while updating massage", ID, userID, err)
+		return nil, err
+	}
+	return nil, nil
 }
 
 // GetAllAppVersions gets all registered versions